// src/types/grpc/user.ts
/**
* This file is auto-generated by nestjs-proto-gen-ts
*/

import { Observable } from 'rxjs';
import { Metadata } from '@grpc/grpc-js';

export namespace user {
    export interface UserService {
        findOne(
            data: UserByEmail,
            metadata?: Metadata,
            ...rest: any[]
        ): Observable<User>;
    }
    export interface UserByEmail {
        email?: string;
    }
    export interface User {
        id?: number;
        email?: string;
        name?: string;
        password?: string;
        token?: string;
        role?: user.Role;
    }
    export interface Role {
        id?: number;
        type?: string;
    }
}